google-search-service/requirements.txt
google-adk
google-generativeai
fastapi
requests
uvicorn
requests
google-cloud-aiplatform
google-api-python-client
deprecated
python-multipart
google-cloud-pubsub